[DISCO-3019] refactor: move api_v1 helper functions to utils & add unit tests
